feat: Add the initial storage implementation
This adds persistence and wires up the command to accept an argument
(the title of a todo) and to append it to the given todo list, if no
list is specified then it will be added to the general todo list.

When a todo is marked as complete, it currently remains in the original
list file but should probably be moved somewhere so it no longer appears
in the UI, this will be added in a later commit.