feat: Add initial automations and overdue labels
This adds some initial automations when the app is launched that will
ensure everything has due dates, and then moves todo items between lists
based on the date they were created, e.g. a todo for tomorrow created
yesterday will appear in the today list when the app is opened.

Additionally this adds an overdue label to todos that are past the due
date.